Inline media controls cut off PiP and fullscreen buttons on cnn.com https://bugs.webkit.org/show_bug.cgi?id=158766 <rdar://problem/24175161>
Patch by Antoine Quint <graouts@apple.com> on 2016-06-14
Reviewed by Dean Jackson.

The display of the picture-in-picture and fullscreen buttons are dependent on the availability
of video tracks through a call to hasVideo(). We need to ensure that the display properties of
both those buttons are updated when the number of video tracks has changed since the controls
may be populated prior to the availability of video tracks.

* Modules/mediacontrols/mediaControlsApple.js:
(Controller.prototype.updateHasVideo):

Identifier: 176854@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202075 268f45cc-cd09-0410-ab3c-d52691b4dbfc