Optimize parseCacheHeader() by using StringView https://bugs.webkit.org/show_bug.cgi?id=158891
Reviewed by Darin Adler.

Source/WebCore:

Optimize parseCacheHeader() and avoid some temporary String allocations
by using StringView. We now strip the whitespaces in the input string
at the beginning of the function, at the same as as we strip the
control characters. We are then able to leverage StringView in the
rest of the function to get substrings without the need for extra
String allocations.

* platform/network/CacheValidation.cpp:
(WebCore::isControlCharacterOrSpace):
(WebCore::trimToNextSeparator):
(WebCore::parseCacheHeader):

Source/WTF:

Add a StringView::find() overload which takes a CharacterMatchFunction
to match the one on String.

* wtf/text/StringView.h:
(WTF::StringView::find):


Identifier: 176964@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202188 268f45cc-cd09-0410-ab3c-d52691b4dbfc