AX: HTML indeterminate IDL attribute not mapped to checkbox value=2 for native checkboxes https://bugs.webkit.org/show_bug.cgi?id=158876 <rdar://problem/26842619>
Reviewed by Joanmarie Diggs.

Source/WebCore:

The indeterminate state was not being reported for native checkboxes.

Also the isIndeterminate() method was relying on whether the appearance changed, which does not happen on Mac, so that
was not being reported correctly. Changed that to check the actual attribute.

Test: accessibility/checkbox-mixed-value.html

* accessibility/AccessibilityNodeObject.cpp:
(WebCore::AccessibilityNodeObject::isIndeterminate):
(WebCore::AccessibilityNodeObject::isPressed):
(WebCore::AccessibilityNodeObject::checkboxOrRadioValue):
* accessibility/AccessibilityObject.cpp:
(WebCore::AccessibilityObject::checkboxOrRadioValue):

LayoutTests:

* accessibility/checkbox-mixed-value-expected.txt: Added.
* accessibility/checkbox-mixed-value.html: Added.


Identifier: 176968@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202194 268f45cc-cd09-0410-ab3c-d52691b4dbfc