[iOS] Typing text into a text field or text area causes screen to scroll down (hiding text entry) https://bugs.webkit.org/show_bug.cgi?id=158970
Reviewed by Ryosuke Niwa.

Source/WebCore:

insertTextWithoutSendingTextEvent() should only reveal the selection up to the main frame on iOS,
since the UI process can zoom and scroll the view to the text input.

Test: fast/forms/ios/typing-in-input-in-iframe.html

* editing/Editor.cpp:
(WebCore::Editor::insertTextWithoutSendingTextEvent):

LayoutTests:

Test that focuses an input in a frame and then types a character, then checks for scrolling.

* fast/forms/ios/typing-in-input-in-iframe-expected.txt: Added.
* fast/forms/ios/typing-in-input-in-iframe.html: Added.

Identifier: 177062@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202295 268f45cc-cd09-0410-ab3c-d52691b4dbfc