REGRESSION: Web Inspector: Text search broken in resources with <CR> https://bugs.webkit.org/show_bug.cgi?id=159110 <rdar://problem/27008485>
Reviewed by Brian Burg.

Source/JavaScriptCore:

* inspector/ContentSearchUtilities.cpp:
(Inspector::ContentSearchUtilities::lineEndings):
The frontend moved to only treated newlines as line endings in
the TextEditor. The backend however was looking for many
different types of line endings (\r\n, \r, \n). This caused
the line endings to ultimately differ between the frontend
and the backend, so the frontend couldn't find the lines that
the backend was claiming search results were on. Change the
backend to only look for \n line endings.

LayoutTests:

* inspector/debugger/searchInContent-linebreaks-expected.txt:
* inspector/debugger/searchInContent-linebreaks.html:
Now that the backend responds with lines that end in \n, this test changes
the number of line results. The frontend interprets this correctly.

Identifier: 177251@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202498 268f45cc-cd09-0410-ab3c-d52691b4dbfc